Remove upper bounds on most dependencies
The cardano-api version we depend on has narrow bounds on cardano-ledger
and plutus-ledger-api versions (which in turn are always only with one
version of plutus-core and plutus-tx compatible).

Consequently, our code base might radically change on each cardano-api
bump, but there is likely no other combination of dependencies that
works than the one the cardano-api package is already constrained on.